+++ packages/typespec-ts/src/index.ts
@@ -6,13 +6,15 @@ import {
   AzureCoreDependencies,
   AzureIdentityDependencies,
   AzurePollingDependencies,
-  DefaultCoreDependencies
+  DefaultCoreDependencies,
+  AzureTestDependencies
 } from "./modular/external-dependencies.js";
 import { clearDirectory } from "./utils/fileSystemUtils.js";
 import { EmitContext, Program } from "@typespec/compiler";
 import { GenerationDirDetail, SdkContext } from "./utils/interfaces.js";
 import {
   CloudSettingHelpers,
+  CreateRecorderHelpers,
   MultipartHelpers,
   PagingHelpers,
   PollingHelpers,
@@ -107,6 +109,7 @@ import { provideSdkTypes } from "./framework/hooks/sdkTypes.js";
 import { transformRLCModel } from "./transform/transform.js";
 import { transformRLCOptions } from "./transform/transfromRLCOptions.js";
 import { emitSamples } from "./modular/emitSamples.js";
+import { emitTests } from "./modular/emitTests.js";
 import { generateCrossLanguageDefinitionFile } from "./utils/crossLanguageDef.js";
 import { getClassicalClientName } from "./modular/helpers/namingHelpers.js";
 
@@ -121,6 +124,12 @@ export async function $onEmit(context: EmitContext) {
   const program: Program = context.program;
   const emitterOptions: EmitterOptions = context.options;
   const dpgContext = await createContextWithDefaultOptions(context);
+
+  // Report any diagnostics from TCGC
+  if (dpgContext.diagnostics?.length > 0) {
+    program.reportDiagnostics(dpgContext.diagnostics);
+  }
+
   // Enrich the dpg context with path detail and common options
   await enrichDpgContext();
   const rlcOptions = dpgContext.rlcOptions ?? {};
@@ -148,10 +157,12 @@ export async function $onEmit(context: EmitContext) {
       ...MultipartHelpers,
       ...CloudSettingHelpers,
       ...XmlHelpers,
+      ...(rlcOptions.generateTest ? CreateRecorderHelpers : {}),
       ...(rlcOptions.enableStorageCompat ? StorageCompatHelpers : {})
     },
     {
       sourcesDir: dpgContext.generationPathDetail?.modularSourcesDir,
+      rootDir: dpgContext.generationPathDetail?.rootDir,
       options: rlcOptions,
       program
     }
@@ -160,7 +171,8 @@ export async function $onEmit(context: EmitContext) {
     ? {
         ...AzurePollingDependencies,
         ...AzureCoreDependencies,
-        ...AzureIdentityDependencies
+        ...AzureIdentityDependencies,
+        ...AzureTestDependencies
       }
     : { ...DefaultCoreDependencies };
   const binder = provideBinder(outputProject, {
@@ -382,7 +394,10 @@ export async function $onEmit(context: EmitContext) {
       }
     }
 
-    binder.resolveAllReferences(modularSourcesRoot);
+    binder.resolveAllReferences(
+      modularSourcesRoot,
+      dpgContext.generationPathDetail?.rootDir
+    );
     if (program.compilerOptions.noEmit || program.hasError()) {
       return;
     }
@@ -454,7 +469,7 @@ export async function $onEmit(context: EmitContext) {
       "test"
     );
     const hasTestFolder = await existsSync(existingTestFolderPath);
-    if (option.azureSdkForJs && option.generateTest === undefined) {
+    if (option.generateTest === undefined) {
       if (hasTestFolder) {
         option.generateTest = false;
       } else {
@@ -487,6 +502,13 @@ export async function $onEmit(context: EmitContext) {
       }
       if (isAzureFlavor) {
         commonBuilders.push(buildEsLintConfig);
+      }
+      if (
+        emitterOptions["generate-test"] === true &&
+        option.azureSdkForJs === true &&
+        emitterOptions["generate-metadata"] === true
+      ) {
+        await emitTests(dpgContext);
       }
       let modularPackageInfo = {};
       if (option.isModularLibrary) {
@@ -536,7 +558,7 @@ export async function $onEmit(context: EmitContext) {
       }
 
       // TODO: need support snippets generation for multi-client cases. https://github.com/Azure/autorest.typescript/issues/3048
-      if (option.generateTest && isAzureFlavor) {
+      if (option.generateTest) {
         for (const subClient of dpgContext.sdkPackage.clients) {
           commonBuilders.push((model) =>
             buildSnippets(model, subClient.name, option.azureSdkForJs)
@@ -633,7 +655,7 @@ export async function $onEmit(context: EmitContext) {
     }
 
     // Generate test relevant files
-    if (option.generateTest && isAzureFlavor && !hasTestFolder) {
+    if (option.generateTest && !hasTestFolder) {
       await emitContentByBuilder(
         program,
         [buildRecordedClientFile, buildSampleTest],