+++ packages/typespec-ts/test/azureIntegration/modelInheritanceEnumDiscriminator.spec.ts
@@ -1,9 +1,10 @@
+import { describe, it, beforeEach, assert } from "vitest";
+
 import EnumDiscriminatorClientFactory, {
   EnumDiscriminatorClient,
   Golden,
   Snake
 } from "./generated/type/model/inheritance/enum-discriminator/src/index.js";
-import { assert } from "chai";
 
 describe("EnumDiscriminatorClient Rest Client", () => {
   let client: EnumDiscriminatorClient;