+++ packages/typespec-ts/test/azureModularIntegration/headersRepeatability.spec.ts
@@ -1,5 +1,6 @@
+import { describe, it, beforeEach, assert } from "vitest";
+
 import { RepeatabilityClient } from "./generated/special-headers/repeatability/src/index.js";
-import { assert } from "chai";
 
 describe("Repeatability Client", () => {
   let client: RepeatabilityClient;