+++ packages/typespec-ts/test/modularIntegration/modelInheritanceNestedDiscriminator.spec.ts
@@ -1,10 +1,11 @@
+import { describe, it, beforeEach, assert } from "vitest";
+
 import {
   GoblinShark,
   NestedDiscriminatorClient,
   Salmon,
   Shark
 } from "./generated/type/model/inheritance/nested-discriminator/src/index.js";
-import { assert } from "chai";
 
 describe("NestedDiscriminatorClient Rest Client", () => {
   let client: NestedDiscriminatorClient;